Complete Project
Need improvement on AI but unfortunately, I had a few personal problems this weekend that consumed a lot of my time and it won't be possible to fix it on time. I could use external plugins to use NavMesh System on 2D World Space or I could import the A* free open source asset on the project.